Model: Deprecated methods (ResNet, CnnVanilla) has been move to an archive folder. FastCnnVanilla and FastResNet has been CnnVanilla and ResNet respectively.
